requirements.txt
fastapi[standard]==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.10.1
python-multipart==0.0.20
orjson==3.11.2
